api/requirements.txt
Flask==3.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
openai>=1.0.0